chore_: move security, version and sentry packages to pkg
