+++ .gitignore
@@ -126,11 +126,11 @@ bindata.go
 migrations.go
 cmd/status-backend/server/endpoints.go
 protocol/messenger_handlers.go
-internal/version/VERSION
-internal/version/GIT_COMMIT
-internal/sentry/SENTRY_CONTEXT_NAME
-internal/sentry/SENTRY_CONTEXT_VERSION
-internal/sentry/SENTRY_PRODUCTION
+pkg/version/VERSION
+pkg/version/GIT_COMMIT
+pkg/sentry/SENTRY_CONTEXT_NAME
+pkg/sentry/SENTRY_CONTEXT_VERSION
+pkg/sentry/SENTRY_PRODUCTION
 
 # Don't ignore generated files in the vendor/ directory
 !vendor/**/*.pb.go
