+++ api/default_networks.go
@@ -5,9 +5,9 @@ import (
 	"strings"
 
 	"github.com/status-im/status-go/api/common"
-	"github.com/status-im/status-go/internal/security"
 	"github.com/status-im/status-go/params"
 	"github.com/status-im/status-go/params/networkhelper"
+	"github.com/status-im/status-go/pkg/security"
 	"github.com/status-im/status-go/protocol/requests"
 )
