+++ params/config.go
@@ -17,11 +17,12 @@ import (
 	"github.com/ethereum/go-ethereum/p2p/discv5"
 	"github.com/ethereum/go-ethereum/params"
 
+	"github.com/status-im/status-go/internal/version"
+
 	"github.com/status-im/status-go/eth-node/crypto"
 	"github.com/status-im/status-go/eth-node/types"
-	"github.com/status-im/status-go/internal/security"
-	"github.com/status-im/status-go/internal/version"
 	"github.com/status-im/status-go/logutils"
+	"github.com/status-im/status-go/pkg/security"
 	"github.com/status-im/status-go/static"
 	wakuv2common "github.com/status-im/status-go/wakuv2/common"
 )
