+++ pkg/security/sensitive_string_test.go
