+++ pkg/sentry/options_test.go
