+++ protocol/requests/create_account.go
@@ -4,8 +4,8 @@ import (
 	"github.com/pkg/errors"
 
 	utils "github.com/status-im/status-go/common"
-	"github.com/status-im/status-go/internal/security"
 	"github.com/status-im/status-go/params"
+	"github.com/status-im/status-go/pkg/security"
 )
 
 var ErrCreateAccountInvalidDisplayName = errors.New("create-account: invalid display name")
