+++ rpc/network/testutil/testutil.go
@@ -4,8 +4,8 @@ import (
 	"github.com/stretchr/testify/require"
 
 	api_common "github.com/status-im/status-go/api/common"
-	"github.com/status-im/status-go/internal/security"
 	"github.com/status-im/status-go/params"
+	"github.com/status-im/status-go/pkg/security"
 )
 
 // Helper function to create a provider
