+++ services/wallet/router/router_test_data.go
@@ -10,8 +10,8 @@ import (
 	"github.com/google/uuid"
 
 	"github.com/status-im/status-go/errors"
-	"github.com/status-im/status-go/internal/security"
 	"github.com/status-im/status-go/params"
+	"github.com/status-im/status-go/pkg/security"
 	walletCommon "github.com/status-im/status-go/services/wallet/common"
 	"github.com/status-im/status-go/services/wallet/requests"
 	"github.com/status-im/status-go/services/wallet/router/fees"
